rename line to record
1. After a quick look in the CODA Febelfin documentation it looks like they
call each line in the CODA file a 'record'. There is a header record and
several data records. It would be beter to stay as close the the jargon
as possible for clarity. It also makes it easier to distinguish between
a line in the file and the conceptual idea of the record on which you
can call other methods.

2. You used `line.line` in the case statement, this is not only
confusing it's also unneeded. You have access to the line of the file in
the block, why would you fetch it from a different object. The `@line`
variable in the `Line` class (see how confusing this gets) is used
inside the scope of the `Line` class and shouldn't be exposed outside of
this object, it's internal state stuff.